add Task 1.08: Added minor changes to README.
add ros2 boilerplate file structure

Deleted the files that will not be used in this project.

added the world file.

created a launch file to launch the world and the model.

modified the CMakeLists.txt and the package.xml according to the new requirements.

modified the CMakeLists.txt

added the screenshot of the gazebo world with the models spawned.

changed the name of the ros package.

Added UML diagram, and Google doc.

add activity diagram and changes to README

add minor changes to README